Factor out common 'errorable' behavior into a module
